# swagger/api.yaml
swagger: '2.0'
info:
  description: The public API for my.move.mil. This is a work in-progress, and is not a final product. We will be continuously updating this site based on feedback.
  version: 0.0.1
  title: my.move.mil
  license:
    name: MIT
    url: https://github.com/transcom/mymove/blob/master/LICENSE.md
basePath: /api/v1
consumes:
  - application/json
produces:
  - application/json
definitions:
...
  UploadPayload:
    type: object
    properties:
      id:
        type: string
        format: uuid
        example: c56a4180-65aa-42ec-a945-5fd21dec0538
      url:
        type: string
        format: uri
        example: https://uploads.domain.test/dir/c56a4180-65aa-42ec-a945-5fd21dec0538
      filename:
        type: string
        format: string
        example: filename.pdf
      content_type:
        type: string
        format: mime-type
        example: application/pdf
      bytes:
        type: integer
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
    required:
      - id
      - url
      - filename
      - content_type
      - bytes
      - created_at
      - updated_at
...
  GBLOC:
    type: string
    description: Office Codes for Transcom offices originating GBLs
    pattern: '^[A-Z]{4}$'
    example: LHNQ
  IndexShipments:
    type: array
    items:
      $ref: '#/definitions/Shipment'
...
  TrafficDistributionList:
    type: object
    properties:
      source_rate_area:
        type: string
        example: US34
      destination_region:
        type: string
        example: '5'
      code_of_service:
        type: string
        example: D
        enum:
          - D
          - '2'
    required:
      - source_rate_area
      - destination_region
      - code_of_service
  Shipment:
    type: object
    properties:
      id:
        type: string
        format: uuid
        example: c56a4180-65aa-42ec-a945-5fd21dec0538
        readOnly: true
      traffic_distribution_list:
        $ref: '#/definitions/TrafficDistributionList'
        readOnly: true
      service_member:
        $ref: '#/definitions/ServiceMember'
        readOnly: true
      actual_pickup_date:
        type: string
        format: date
        example: '2018-04-21'
        x-nullable: true
      actual_pack_date:
        type: string
        format: date
        example: '2018-04-21'
        x-nullable: true
      actual_delivery_date:
        type: string
        format: date
        example: '2018-04-27'
        x-nullable: true
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
      source_gbloc:
        $ref: '#/definitions/GBLOC'
        readOnly: true
      destination_gbloc:
        $ref: '#/definitions/GBLOC'
        readOnly: true
      gbl_number:
        type: string
        example: 'KKFA9999999'
        x-nullable: true
        title: GBL Number
      market:
        $ref: '#/definitions/ShipmentMarket'
        readOnly: true
        x-nullable: true
      book_date:
        type: string
        format: date
        example: '2018-03-15'
      requested_pickup_date:
        type: string
        format: date
        example: '2018-04-19'
      original_delivery_date:
        type: string
        format: date
        example: '2018-04-26'
        readOnly: true
        x-nullable: true
      original_pack_date:
        type: string
        format: date
        example: '2018-04-26'
        readOnly: true
        x-nullable: true
      move:
        $ref: '#/definitions/Move'
        x-nullable: true
      status:
        $ref: '#/definitions/ShipmentStatus'
        readOnly: true
      estimated_pack_days:
        type: integer
        description: Estimated days packing
        minimum: 0
        example: 2
        x-nullable: true
      estimated_transit_days:
        type: integer
        description: Estimated days in transit
        minimum: 0
        example: 30
        x-nullable: true
      pickup_address:
        $ref: '#/definitions/Address'
        x-nullable: true
      has_secondary_pickup_address:
        type: boolean
        title: 'Do you have household goods at any other pickup location?'
      secondary_pickup_address:
        $ref: '#/definitions/Address'
        x-nullable: true
      has_delivery_address:
        type: boolean
        title: 'Do you know the delivery address at destination yet?'
      delivery_address:
        $ref: '#/definitions/Address'
        x-nullable: true
      has_partial_sit_delivery_address:
        type: boolean
      partial_sit_delivery_address:
        $ref: '#/definitions/Address'
        x-nullable: true
      weight_estimate:
        type: integer
        description: Estimated weight of the shipment in lbs
        minimum: 0
        example: 10000
        x-nullable: true
        x-formatting: weight
        title: Weight Estimate
      progear_weight_estimate:
        type: integer
        description: Estimated weight of the progear in lbs
        minimum: 0
        example: 500
        x-nullable: true
        x-formatting: weight
        title: Progear Weight Estimate
      spouse_progear_weight_estimate:
        type: integer
        description: Estimated weight of the spouse progear in lbs
        minimum: 0
        example: 200
        x-nullable: true
        x-formatting: weight
        title: Spouse Progear Weight Estimate
      net_weight:
        type: integer
        description: Actual net weight of the shipment in lbs
        minimum: 0
        example: 10000
        x-nullable: true
        x-formatting: weight
        title: Net
      gross_weight:
        type: integer
        description: Gross weight of the shipment in lbs
        minimum: 0
        example: 10000
        x-nullable: true
        x-formatting: weight
        title: Gross
      tare_weight:
        type: integer
        description: Tare weight of the shipment in lbs
        minimum: 0
        example: 10000
        x-nullable: true
        x-formatting: weight
        title: Tare
      pm_survey_conducted_date:
        type: string
        format: date
        example: '2018-04-26'
        title: 'Conducted'
        x-nullable: true
      pm_survey_planned_pack_date:
        type: string
        format: date
        example: '2018-04-26'
        title: 'Planned pack date'
        x-nullable: true
      pm_survey_planned_pickup_date:
        type: string
        format: date
        example: '2018-04-26'
        title: 'Planned pickup date'
        x-nullable: true
      pm_survey_planned_delivery_date:
        type: string
        format: date
        example: '2018-04-26'
        title: 'Planned delivery date'
        x-nullable: true
      pm_survey_weight_estimate:
        type: integer
        minimum: 0
        example: 10000
        title: Weight
        x-nullable: true
        x-formatting: weight
      pm_survey_progear_weight_estimate:
        type: integer
        minimum: 0
        example: 10000
        title: Pro-Gear weight
        x-nullable: true
        x-formatting: weight
      pm_survey_spouse_progear_weight_estimate:
        type: integer
        minimum: 0
        example: 10000
        title: Spouse's pro-gear
        x-nullable: true
        x-formatting: weight
      pm_survey_notes:
        type: string
        example: This document is good to go!
        format: textarea
        x-nullable: true
        title: Notes about dates
      pm_survey_method:
        type: string
        title: Survey method
        enum:
          - IN_PERSON
          - PHONE
          - VIDEO
        x-display-value:
          IN_PERSON: In person
          PHONE: Phone
          VIDEO: Video
...
  ShipmentMarket:
    type: string
    description: One of the possible 'Markets' for a shipment
    example: dHHG
    enum:
      - dHHG
      - iHHG
      - iUB
    x-display-value:
      dHHG: Domestic HHG
      iHHG: International HHG
      iUB: International unaccompanied baggage
...
  InvalidRequestResponsePayload:
    type: object
    properties:
      errors:
        type: object
        additionalProperties:
          type: string
paths:
...
  /uploads:
    post:
      summary: Create a new upload
      description: Uploads represent a single digital file, such as a JPEG or PDF.
      operationId: createUpload
      tags:
        - uploads
      consumes:
        - multipart/form-data
      parameters:
        - in: query
          name: documentId
          type: string
          format: uuid
          required: false
          description: UUID of the document to add an upload to
        - in: formData
          name: file
          type: file
          description: The file to upload.
          required: true
      responses:
        201:
          description: created upload
          schema:
            $ref: '#/definitions/UploadPayload'
        400:
          description: invalid request
          schema:
            $ref: '#/definitions/InvalidRequestResponsePayload'
        403:
          description: not authorized
        404:
          description: not found
        500:
          description: server error
    delete:
      summary: Deletes a collection of uploads
      description: Uploads represent a single digital file, such as a JPEG or PDF.
      operationId: deleteUploads
      tags:
        - uploads
      parameters:
        - in: query
          name: uploadIds
          type: array
          items:
            type: string
            format: uuid
          required: true
          description: Array of UUIDs to be deleted
      responses:
        204:
          description: deleted
        400:
          description: invalid request
          schema:
            $ref: '#/definitions/InvalidRequestResponsePayload'
        403:
          description: not authorized
        404:
          description: not found
        500:
          description: server error
  /shipments:
    get:
      summary: Gets visible shipments
      description: Allows a user to retrieve a list of shipments. The shipments are restricted to those they have authorization to view
      operationId: indexShipments
      tags:
        - shipments
      parameters:
        - name: status
          in: query
          type: array
          items:
            type: string
            enum: *SHIPMENTSTATUS
          collectionFormat: csv
          description: Restrict the list to shipments with matching ShipmentStatus
        - name: order_by
          in: query
          type: string
          description: Sort the returned list according to this ordering
          enum:
            - PICKUP_DATE_ASC
            - PICKUP_DATE_DESC
            - DELIVERY_DATE_ASC
            - DELIVERY_DATE_DESC
        - name: limit
          in: query
          type: integer
          description: maximum number of records to return
          minimum: 1
          default: 25
        - name: offset
          in: query
          type: integer
          description: offset into records to select from. Actual offset is this value multiplied by limit.
          minimum: 1
          default: 1
      responses:
        200:
          description: returns an array of Shipments
          schema:
            $ref: '#/definitions/IndexShipments'
        400:
          description: invalid request
        401:
          description: must be authenticated to use this endpoint
        403:
          description: not authorized to list these shipments
        422:
          description: cannot process request with given information
        500:
          description: server error
...